pyproject.toml
[project]
name = "social-media-automation"
version = "0.1.0"
description = "Multi-persona social media automation agent powered by LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.45.0",
    "matplotlib>=3.9.0",
    "pillow>=10.4.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
